pyproject.toml
[project]
name = "webmaster-domain-tool"
version = "1.0.0"
description = "Comprehensive domain analysis tool for webmasters"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Webmaster Tools", email = "tools@example.com" }
]
license = { text = "MIT" }
keywords = ["dns", "ssl", "security", "webmaster", "domain", "analysis"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Networking",
    "Topic :: Security",
]

dependencies = [
    "dnspython>=2.8.0",
    "httpx>=0.28.1",
    "cryptography>=46.0.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "tomli>=2.3.0; python_version < '3.11'",
    "tomli_w>=1.0.0",
    "python-whois>=0.9.6",
    "flet>=0.28.3",
    "flet-desktop>=0.28.3",
    "beautifulsoup4>=4.12.0",
    "html5lib>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "black>=25.11.0",
    "ruff>=0.14.1",
    "mypy>=1.18.2",
]

[project.urls]
Homepage = "https://github.com/orgoj/webmaster-domain-tool"
Repository = "https://github.com/orgoj/webmaster-domain-tool"
Issues = "https://github.com/orgoj/webmaster-domain-tool/issues"

[project.scripts]
webmaster-domain-tool = "webmaster_domain_tool.cli:main"
wdt = "webmaster_domain_tool.cli:main"
wdt-app = "webmaster_domain_tool.flet_app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/webmaster_domain_tool"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B904",  # raise from err - TODO: fix these properly
    "B008",  # function call in defaults (typer pattern)
    "B007",  # unused loop variable
    "C401",  # unnecessary generator
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--verbose --cov=webmaster_domain_tool --cov-report=term-missing"

[dependency-groups]
dev = [
    "isort>=7.0.0",
    "pre-commit>=4.4.0",
    "pyright>=1.1.407",
    "pytest-cov>=7.0.0",
]
